requirements.txt
Django == 1.11.12
Pillow == 3.3.1
Top == 1.0.0
PyMySQL == 0.8.1